move to generateTypeScript.ts
